Implemented Analyzer Clustering
