+++ conf/application.conf
@@ -47,7 +47,7 @@ play.i18n.langs = [ "en" ]
 # -----
 akka {
   actor {
-    provider = "akka.remote.RemoteActorRefProvider"
+    provider = "akka.cluster.ClusterActorRefProvider"
   }
 
   remote {
@@ -56,76 +56,41 @@ akka {
       port = 2550
     }
   }
-}
 
-# Settings for the ClusterClient
-akka.cluster.client {
-  # Actor paths of the ClusterReceptionist actors on the servers (cluster nodes)
-  # that the client will try to contact initially. It is mandatory to specify
-  # at least one initial contact.
-  # Comma separated full actor paths defined by a string on the form of
-  # "akka.tcp://system@hostname:port/system/receptionist"
-  initial-contacts = []
+  cluster {
+    seed-nodes = [
+      "akka.tcp://application@127.0.0.1:2550",
+      "akka.tcp://application@127.0.0.1:2551"]
 
-  # Interval at which the client retries to establish contact with one of
-  # ClusterReceptionist on the servers (cluster nodes)
-  establishing-get-contacts-interval = 3s
+    auto-down-unreachable-after = 5 seconds
 
-  # Interval at which the client will ask the ClusterReceptionist for
-  # new contact points to be used for next reconnect.
-  refresh-contacts-interval = 60s
-
-  # How often failure detection heartbeat messages should be sent
-  heartbeat-interval = 1s
-
-  # Number of potentially lost/delayed heartbeats that will be
-  # accepted before considering it to be an anomaly.
-  # The ClusterClient is using the akka.remote.DeadlineFailureDetector, which
-  # will trigger if there are no heartbeats within the duration
-  # heartbeat-interval + acceptable-heartbeat-pause, i.e. 15 seconds with
-  # the default settings.
-  acceptable-heartbeat-pause = 3s
-
-  # If connection to the receptionist is not established the client will buffer
-  # this number of messages and deliver them the connection is established.
-  # When the buffer is full old messages will be dropped when new messages are sent
-  # via the client. Use 0 to disable buffering, i.e. messages will be dropped
-  # immediately if the location of the singleton is unknown.
-  # Maximum allowed buffer size is 10000.
-  buffer-size = 1000
-}
-
-# Settings for the ClusterReceptionistExtension
-akka.contrib.cluster.receptionist {
-  # Actor name of the ClusterReceptionist actor, /user/receptionist
-  name = receptionist
-
-  # Start the receptionist on members tagged with this role.
-  # All members are used if undefined or empty.
-  role = ""
-
-  # The receptionist will send this number of contact points to the client
-  number-of-contacts = 3
-
-  # The actor that tunnel response messages to the client will be stopped
-  # after this time of inactivity.
-  response-tunnel-receive-timeout = 500 milliseconds
+    roles = [monitor]
+  }
 }
 
 solar-farm-monitor {
   analyzer {
-    contact-points = [
-      "akka.tcp://solar-farm-analyzer@127.0.0.1:2551"
-      "akka.tcp://solar-farm-analyzer@127.0.0.1:2552"]
+    cluster.role = "analyzer"
     inspection-interval = 1 second
     inspection-response-timeout = 2 seconds
     measurement-interval = 1 second
     measurement-response-timeout = 2 seconds
     error-notification-interval = 1 second
-    unstable-timeout = 5 minutes
-    initial-connection-attempt-timeout = 2 seconds
+    unreachable-timeout = 5 minutes
   }
 }
+
+# Cluster Metrics
+# ---------------
+# Disable legacy metrics in akka-cluster.
+akka.cluster.metrics.enabled=off
+
+# Enable metrics extension in akka-cluster-metrics.
+# akka.extensions=["akka.cluster.metrics.ClusterMetricsExtension"]
+
+# Sigar native library extract location during tests.
+# Note: use per-jvm-instance folder when running multiple jvm on one host.
+akka.cluster.metrics.native-library-extract-folder=${user.dir}/target/native
 
 # Log
 # -----
@@ -142,6 +107,6 @@ akka {
     fsm = on
     event-stream = off
   }
-  cluster.log-info = off
+  cluster.log-info = on
   remote.log-remote-lifecycle-events = on
 }
