+++ modules/analyzer/src/main/resources/application.conf
@@ -2,8 +2,6 @@ akka {
 
   actor.provider = "akka.cluster.ClusterActorRefProvider"
 
-  extensions = ["akka.contrib.pattern.ClusterReceptionistExtension"]
-
   remote {
     netty.tcp {
       hostname = "127.0.0.1"
@@ -13,17 +11,12 @@ akka {
 
   cluster {
     seed-nodes = [
-      "akka.tcp://"${solar-farm-analyzer.cluster.name}"@127.0.0.1:2551",
-      "akka.tcp://"${solar-farm-analyzer.cluster.name}"@127.0.0.1:2552"]
+      "akka.tcp://"${solar-farm-analyzer.cluster.name}"@127.0.0.1:2550",
+      "akka.tcp://"${solar-farm-analyzer.cluster.name}"@127.0.0.1:2551"]
 
-    auto-down-unreachable-after = 5 seconds
-  }
+    roles = [analyzer]
 
-  contrib.cluster.client {
-    mailbox {
-      mailbox-type = "akka.dispatch.UnboundedDequeBasedMailbox"
-      stash-capacity = 1000
-    }
+    akka.cluster.auto-down-unreachable-after = 10s
   }
 }
 
@@ -38,3 +31,28 @@ akka.cluster.metrics.enabled=off
 # Sigar native library extract location during tests.
 # Note: use per-jvm-instance folder when running multiple jvm on one host.
 akka.cluster.metrics.native-library-extract-folder=${user.dir}/target/native
+
+akka.actor.deployment {
+
+  /singleton/analysis-supervisor/inspection-supervisor/calculation-supervisor/sum-calculator-router {
+    router = round-robin-pool
+    nr-of-instances = 100
+    cluster {
+      enabled = on
+      max-nr-of-instances-per-node = 4
+      allow-local-routees = off
+      use-role = analyzer
+    }
+  }
+
+  /singleton/analysis-supervisor/inspection-supervisor/calculation-supervisor/inspector-router {
+    router = round-robin-pool
+    nr-of-instances = 100
+    cluster {
+      enabled = on
+      max-nr-of-instances-per-node = 4
+      allow-local-routees = off
+      use-role = analyzer
+    }
+  }
+}
