--- modules/analyzer/src/main/scala/com/example/analyer/actors/inspection/CalculationSupervisor.scala
@@ -1,30 +1,0 @@
-package com.example.analyer.actors.inspection
-
-import akka.actor.SupervisorStrategy.Restart
-import akka.actor._
-
-import scala.concurrent.duration._
-
-class CalculationSupervisor extends Actor with ActorLogging {
-
-  override val supervisorStrategy: SupervisorStrategy =
-    AllForOneStrategy(maxNrOfRetries = 100, withinTimeRange = 1 minute) {
-      case _ => Restart
-    }
-
-  def receive = Actor.emptyBehavior
-
-  def createSumCalculator() =
-    context.actorOf(SumCalculator.props(createMeanCalculator()), "sum-calculator")
-
-  def createMeanCalculator() =
-    context.actorOf(LowerLimitCalculator.props(createInspector()), "mean-calculator")
-
-  def createInspector() =
-    context.actorOf(Props[Inspector], "inspector")
-
-  override def preStart() = {
-    createSumCalculator()
-  }
-
-}
