+++ modules/backend-config/src/main/resources/reference.conf
@@ -20,10 +20,11 @@ akka {
 # Analyzer
 # -----
 solar-farm-analyzer {
-  cluster.name = "solar-farm-analyzer"
+  cluster.name = "application"
   mqtt.client.id-prefix = ${solar-farm-analyzer.cluster.name}
   mqtt.topic.root = ${solar-farm-simulator.mqtt.topic.root}
-  inspector.alert-threshold-per = 20
+  inspecting-timeout = 5 seconds
+  inspector.alert-threshold-per = 30
   buffer {
     ghost-collection-initial-delay = 500 milliseconds
     ghost-collection-interval = 10 seconds
